V3.71 (#356)
* incrementing SEMVER to v3.71

* augmenting the skeleton structure to best known practice

* embellishing the oneparam library standard header

* cleaning up orphaned code

* cleanup on the IEEE-754 quire isle

* adding documentation to the quire number type for future research

* WIP: adding error free transform operators

* WIP: faithful skeleton

* WIP: faithful addition

* WIP: implementation of compensated addition twosum style

* WIP: adding compensated evaluations for multiplication and division

* WIP: progress to supporting cfloat in faithful arithmetic

* WIP: experiment to include classic floats in the faithful shim

* expanding converting constructor to support fp8/fp16/fp32/fp64 conversions

* WIP: twosum and fma algorithm updates

* WIP: enable custom fma for cfloats smaller than 40bits

* removing warning from Clang/GCC builds

* fix code documentation to reflect current design

* renaming lns2b to dbns

* WIP: generalizing the cfloat conversion by virtue of the blocktriple

* WIP: generalizing cfloat conversion

* WIP: adding a blockfraction to deliver a binary to decimal string conversion facility

* WIP: declaring blockfraction to be unsigned

* compilation fix for clang and gcc

* WIP: reverse engineering the Dragon4 algorithm

* WIP: dragon1 test case

* WIP: debugging the dynamic stopping condition

* Dragon1 example with correct exit

* WIP: highlighting the subtleties of decimal conversion

* numeric_limits fix for faithful type

---------

+++ include/universal/native/constexpr754.hpp
@@ -56,7 +56,7 @@ inline constexpr void extractFields(long double value, bool& s, uint64_t& rawExp
 #ifdef LONG_DOUBLE_DOWNCAST
 template<>
 inline constexpr void extractFields(long double value, bool& s, uint64_t& rawExponentBits, uint64_t& rawFractionBits) noexcept {
-	double d = static_cast<long double>(value);
+	double d = static_cast<double>(value);
 	uint64_t bc = std::bit_cast<uint64_t, double>(d);
 	s = (ieee754_parameter<double>::smask & bc);
 	rawExponentBits = (ieee754_parameter<double>::emask & bc) >> ieee754_parameter<double>::fbits;
